+++ api/settings.py
@@ -19,9 +19,9 @@ from constants import (
     STATIC_ROOT,
     BASE_URL,
     VUE_ROOT,
+    get_database_configuration,
 )
 from datetime import timedelta
-from get_database_configuration import get_database_configuration
 
 # Build paths inside the project like this: os.path.join(BASE_DIR, ...)
 BASE_DIR = os.path.dirname(os.path.dirname(os.path.abspath(__file__)))
